Component cleanup
Removed unused components Login, Logout, Transactions, CreateAccount.
Removed references to these files and routes from App.js, as well as the links in the navigation bar.

Removed Deposit.jsx and Withdrawal.jsx

Updated App.js to render the BankCard when routing to Withdrawal or Deposit, and to send the appropriate transactionType prop.